requirements.txt
appdirs==1.4.4
bidict==0.21.2
black==21.5b2
certifi==2020.6.20
chardet==3.0.4
click==8.0.1
dataclasses==0.8
greenlet==1.1.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
irc==18.0.0
isodate==0.6.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.logging==3.0.0
jaraco.stream==3.0.0
jaraco.text==3.2.0
more-itertools==8.4.0
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.8.1
psycopg2-binary
pydub==0.25.1
pyright==0.0.13
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2020.1
regex==2021.4.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.4.20
tempora==3.0.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.25.9
websocket-client==1.1.0
youtube-dl
zipp==3.1.0
